requirement.txt
langchain>=0.1.0
langchain-community>=0.1.0
langchain-core>=0.1.0
dashscope>=1.14.0
requests>=2.31.0
python-dotenv>=1.0.0